refactored campaign game into different files to make it easier to find components
