+++ frontend/src/campaign/campaignView.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import * as PropTypes from 'prop-types';
 
 // eslint-disable-next-line no-unused-vars
 import Popover from 'react-bootstrap/Popover';
@@ -12,529 +11,11 @@ import { project_features_and_create_svg_paths } from '../common';
 import { MapPath } from '../UILibrary/components';
 import './campaign.scss';
 
+import { Speech, get_country_prop, get_default_proposal } from './Speech';
+import Feedback from './Feedback';
+import Results from './Results';
+import CountrySelectorPopup from './CountrySelectorPopup';
 import IntroView from '../chooseAdventure/introView';
-
-// NOTE(ra): this is a person icon path that I just bounced out of Illustrator and am hardcoding
-// here for expediency. We should really put this in an SVG file and get it into the project
-// in a more reasonable way!
-// eslint-disable-next-line max-len
-const PERSON_ICON_PATH = 'M437.02,330.98c-27.883-27.882-61.071-48.523-97.281-61.018C378.521,243.251,404,198.548,404,148 C404,66.393,337.607,0,256,0S108,66.393,108,148c0,50.548,25.479,95.251,64.262,121.962 c-36.21,12.495-69.398,33.136-97.281,61.018C26.629,379.333,0,443.62,0,512h40c0-119.103,96.897-216,216-216s216,96.897,216,216 h40C512,443.62,485.371,379.333,437.02,330.98z M256,256c-59.551,0-108-48.448-108-108S196.449,40,256,40 c59.551,0,108,48.448,108,108S315.551,256,256,256z';
-
-const get_default_proposal = (topic_names) => {
-    const proposal = {};
-    topic_names.forEach((topic) => {
-        proposal[topic] = 3;
-    });
-    return proposal;
-};
-
-// const all_topics = [
-//     'Health services',
-//     'Education',
-//     'Water and sanitation',
-//     'Roads and bridges',
-//     'Electricity',
-//     'Equal rights for women',
-//     'Improving living standards for the poor',
-//     'Creating jobs',
-//     'Fighting corruption',
-//     'Reducing violent community conflict',
-// ];
-
-const COUNTRIES = [
-    {
-        name: 'Botswana',
-        ISO: 'BWA',
-        population: '2.254 million',
-        difficulty: 'Easy',
-        topicNames: [
-            'Education',
-            'Creating jobs',
-            'Fighting corruption',
-        ],
-        supportThreshold: 0.5,
-        max_priority_points: 12,
-    },
-    {
-        name: 'Kenya',
-        ISO: 'KEN',
-        population: '51.39 million',
-        difficulty: 'Medium',
-        topicNames: [
-            'Health services',
-            'Education',
-            'Water and sanitation',
-            'Equal rights for women',
-            'Creating jobs',
-            'Fighting corruption',
-        ],
-        supportThreshold: 2,
-        max_priority_points: 22,
-    },
-    {
-        name: 'South Africa',
-        ISO: 'ZAF',
-        population: '57.78 million',
-        difficulty: 'Hard',
-        topicNames: [
-            'Health services',
-            'Education',
-            'Water and sanitation',
-            'Roads and bridges',
-            'Electricity',
-            'Equal rights for women',
-            'Improving living standards for the poor',
-            'Creating jobs',
-            'Fighting corruption',
-            'Reducing violent community conflict',
-        ],
-        supportThreshold: 12,
-        max_priority_points: 35,
-    },
-];
-
-function get_country_prop(country_name, prop_name) {
-    // Each country has its own set of topicNames
-    const country = COUNTRIES.filter((obj) => {
-        return obj.name === country_name;
-    })[0];
-    return country[prop_name];
-}
-
-
-class Speech extends React.Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            speechProposal: this.props.speechProposal,
-            result: 0,
-            total: Object.keys(this.props.speechProposal).reduce((acc, topic) => {
-                return acc + this.props.speechProposal[topic];
-            }, 0),
-        };
-        this.difference_threshold = get_country_prop(this.props.countryName, 'supportThreshold');
-        this.max_priority_points = get_country_prop(this.props.countryName, 'max_priority_points');
-    }
-
-    /**
-     * Resets the speech such that every topic has a value of 1.
-     * Used when component mounts and upon onClick of a button
-     */
-    resetSpeech = () => {
-        this.setState({
-            speechProposal: get_default_proposal(this.props.topicNames),
-            total: 30,
-        }, () => {
-            this.setState({ result: this.countSupporters() });
-        });
-    };
-
-    componentDidMount() {
-        if (this.props.canReset) {
-            this.countSupporters(); // in the case that user changed nothing
-        }
-    }
-
-    generateStory() {
-        if (this.props.round === 1) {
-            return 'You just started your campaign and we want you to set out '
-                    + 'your policies so that we can conduct polls to gauge the initial reaction.';
-        }
-        if (this.props.round === 2) {
-            return 'We can do two more polls before elections.';
-        }
-        return 'This is the final policies that people will see in the election';
-    }
-
-    componentDidUpdate(prevProps) {
-        if (prevProps.countryName !== this.props.countryName) {
-            this.resetSpeech();
-        }
-        if (prevProps.countryName !== this.props.countryName
-            || prevProps.population !== this.props.population) {
-            this.setState({
-                result: this.countSupporters(),
-            });
-        }
-    }
-
-    /**
-     * Handles when the slider changes by changing the state of what the maximum values should
-     * be for each category and updates the number of supporters
-     * @param e The event that is triggered, use e.target.value to get the value of the slider
-     * @param topic Tells which topic the slider belongs to so that it updates the speech
-     */
-    handleButtonOnChange = (e, topic) => {
-        const newVal = parseInt(e.target.value);
-        const newProposal = this.props.speechProposal;
-        const oldVal = newProposal[topic];
-        if (this.state.total + newVal - oldVal <= this.max_priority_points) {
-            newProposal[topic] = newVal;
-            this.setState({
-                speechProposal: newProposal,
-                total: this.state.total + newVal - oldVal,
-                result: this.countSupporters(),
-            });
-        }
-    };
-
-    /**
-     * Determines whether a citizen will support you based on your speech.
-     * @returns {number} Number of people that will support you
-     */
-    countSupporters = () => {
-        let count = 0;
-        Object.keys(this.props.population).forEach((province) => {
-            let numSupporters = 0;
-            this.props.population[province]['citizens'].forEach((citizen) => {
-                let difference_score = 0;
-                for (const topic of Object.keys(this.state.speechProposal)) {
-                    if (this.state.speechProposal[topic] < citizen['traits'][topic]) {
-                        difference_score += (citizen['traits'][topic]
-                            - this.state.speechProposal[topic]) ** 2;
-                    }
-                }
-
-                if (difference_score > this.difference_threshold) {
-                    citizen.will_support = false;
-                } else {
-                    citizen.will_support = true;
-                    numSupporters += 1;
-                    count++;
-                }
-            });
-            this.props.population[province]['totalSupporters'] = numSupporters;
-        });
-
-        this.props.updatePopulation(this.props.population);
-        return count;
-    };
-
-    priorityPoint() {
-        if (this.max_priority_points - this.state.total > 0) {
-            return 'You can prioritize more things.';
-        }
-        return 'You need to de-prioritize others first';
-    }
-
-    render() {
-        const topics = this.props.topicNames.map((topic, key) => (
-            <div key={key} className='speech-option'>
-                <div className='speech-option_label'>
-                    {topic}
-                </div>
-                <div className='speech-option_btns'>
-                    {[...Array(5).keys()].map((score, j) => (
-                        <input className='speech-radio' type='radio' name={topic} key={j}
-                            id={'inlineRadio' + score + 1} value={score + 1}
-                            checked={this.state.speechProposal[topic] === score + 1}
-                            onChange={(e) => this.handleButtonOnChange(e, topic)}/>
-                        // </div>
-                    ))}
-                </div>
-            </div>
-        ));
-
-        return (
-            <div className="row w-100"><div className='col'>
-                <div className='speech-context'>
-                    <p className='speech-context_count'>
-                        {this.generateStory()}
-                    </p>
-                    <div className='speech-context_points'>
-                        {this.priorityPoint()}
-                    </div>
-                </div>
-                <div className='speech-options'>
-                    <div className='speech-option-desc'>
-                        <span>Low priority</span>
-                        <span>High priority</span>
-                    </div>
-                    {topics}
-                </div>
-                <div className='reset_button'>
-                    <button
-                        className='campaign-btn speech-btn'
-                        onClick={this.props.submitPriorities}
-                    >
-                        Submit
-                    </button>
-                </div>
-            </div></div>
-        );
-    }
-}
-Speech.propTypes = {
-    population: PropTypes.object,
-    countryName: PropTypes.string,
-    updatePopulation: PropTypes.func,
-    submitPriorities: PropTypes.func,
-    speechProposal: PropTypes.object,
-    topicNames: PropTypes.array,
-    canReset: PropTypes.bool,
-    round: PropTypes.number,
-};
-
-class Feedback extends React.Component {
-    constructor(props) {
-        super(props);
-    }
-
-    render() {
-        const { results, clickedProvince } = this.props;
-        let citizenReactions;
-        if (clickedProvince) {
-            const sample = results[clickedProvince]['citizens'].slice(0, 100);
-            citizenReactions = sample.map((citizen, k) => (
-                <Citizen
-                    key={k}
-                    data={citizen}
-                    title={`Citizen of ${citizen['province']}`}
-                    generateDescription={this.props.generateDescription}
-                />
-            ));
-        }
-        const description = (<div>
-            {clickedProvince
-                ? (<strong>
-                    Round {this.props.round - 1} results for {clickedProvince}
-                </strong>)
-                : (<></>)}
-        </div>);
-
-        return (
-            <div className='feedback'>
-                <div className='feedback-results'>
-                    <p>
-                    Click on a province, and then mouse over citizens to see their reactions to your
-                    proposal.
-                    </p>
-
-                    {description}
-                    <div className='feedback-pop'>{citizenReactions}</div>
-                </div>
-                <button className='campaign-btn' onClick={this.props.nextRound}>
-                    Next Round
-                </button>
-            </div>
-        );
-    }
-}
-Feedback.propTypes = {
-    clickedProvince: PropTypes.string,
-    generateDescription: PropTypes.func,
-    results: PropTypes.object,
-    round: PropTypes.number,
-    nextRound: PropTypes.func,
-};
-
-
-class Results extends React.Component {
-    constructor(props) {
-        super(props);
-        this.map_height = 500;
-        this.map_width = 500;
-    }
-
-    render() {
-        const resultsData = this.props.provinceData;
-        if (!resultsData) {
-            return (<></>);
-        }
-
-        // TODO: possibly refactor this further since it is similar to CampaignView
-        const sample = [];
-        Object.values(this.props.provinceData).forEach((province) => {
-            const citizens = province['citizens'];
-            sample.push(...citizens.slice(0, Math.round(citizens.length * 0.25)));
-        });
-        const citizens = sample.map((citizen, k) => (
-            <Citizen
-                key={k}
-                data={citizen}
-                title={`Citizen of ${citizen['province']}`}
-                generateDescription={this.props.generateDescription}
-            />
-        ));
-
-        return (
-            <div>
-                <table border='1' className={'resultTable'}>
-                    <tbody>
-                        <tr>
-                            <th>Province Name</th>
-                            <th>Number of Supporters</th>
-                            <th>Number of People</th>
-                            <th>Percentage of Votes</th>
-                        </tr>
-                        {Object.keys(resultsData).map((province, k) => (
-                            (
-                                province !== 'countryTotal'
-                                && province !== 'countrySupporters'
-                                && province !== 'countryName'
-                            )
-                            && <tr key={k}>
-                                <td>{province}</td>
-                                <td>{resultsData[province].totalSupporters}</td>
-                                <td>{resultsData[province].citizens.length}</td>
-                                <td>{Math.round((resultsData[province].totalSupporters
-                                    / resultsData[province].citizens.length) * 100)}%</td>
-                            </tr>
-                        ))}
-                        <tr className={'countryResult'}>
-                            <th>{this.props.countryName}</th>
-                            <th>{this.props.countryData.totalSupport}</th>
-                            <th>{this.props.countryData.totalPopulation}</th>
-                            <th>{Math.round((this.props.countryData.totalSupport
-                                                / this.props.countryData.totalPopulation) * 100)}%
-                            </th>
-                        </tr>
-                    </tbody>
-                </table>
-                <div className='campaign-result_graphics'>
-                    <svg
-                        height={this.map_height}
-                        width={this.map_width}
-                        id='content'
-                        className='result-map'
-                    >
-                        {this.props.mapData.map((province, i) => {
-                            let countryFill;
-                            if (province.name) {
-                                countryFill = resultsData[province.name].totalSupporters
-                                / resultsData[province.name].citizens.length > 0.5
-                                    ? '#5abf5a' : '#db5653';
-                            } else {
-                                countryFill = '#F6F4D2';
-                            }
-
-                            return <MapPath
-                                key={i}
-                                path={province.svg_path}
-                                id={province.postal}
-                                fill={countryFill}
-                                stroke='black'
-                                strokeWidth='1'
-                                useColorTransition={false}
-                            />;
-                        })}
-                    </svg>
-                    <div className='result-population'>
-                        <div className='result-population_header'>
-                            Results for sample population of size {sample.length}
-                        </div>
-                        <div className='result-population_svg'>
-                            {citizens}
-                        </div>
-                    </div>
-
-                </div>
-            </div>
-        );
-    }
-}
-Results.propTypes = {
-    provinceData: PropTypes.object,
-    countryData: PropTypes.object,
-    countryName: PropTypes.string,
-    mapData: PropTypes.array,
-    generateDescription: PropTypes.func,
-};
-
-/**
- * Component for displaying citizen information
- *
- * The fill indicates the citizen's stance on the budget,
- * and hovering over the component displays the citizen's traits
- */
-export class Citizen extends React.Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            show: false,
-        };
-    }
-
-    render() {
-        const description = (
-            <Popover id='popover-basic'>
-                <Popover.Title>
-                    {this.props.title}
-                </Popover.Title>
-                <Popover.Content>
-                    {this.props.generateDescription(this.props.data)}
-                </Popover.Content>
-            </Popover>
-        );
-        return (
-            <OverlayTrigger
-                overlay={description}
-                placement='right'
-            >
-                <svg
-                    className='citizen'
-                    height='20'
-                    width='20'
-                    viewBox="0 0 512 512"
-                    fill={this.props.data.will_support ? '#5abf5a' : '#db5653'}
-                >
-                    <path d={PERSON_ICON_PATH} />
-                </svg>
-            </OverlayTrigger>
-        );
-    }
-}
-Citizen.propTypes = {
-    data: PropTypes.object,
-    title: PropTypes.string,
-    generateDescription: PropTypes.func,
-};
-
-class CountrySelectorPopup extends React.Component {
-    selectCountryClosePopup = (country) => {
-        this.props.changeCountry(country);
-        this.props.closePopup();
-    };
-
-    render() {
-        return (
-            <div className='country-selector'>
-                <h3>Select a country</h3>
-                <div className='row w-100'>
-                    {COUNTRIES.map((country, key) => (
-                        <div key={key} className='col-4'>
-                            <div className='card' >
-                                <div className='card-header'>
-                                    <h5>{country.name}</h5>
-                                </div>
-                                <div className='card-body'>
-                                    <table className='table'><tbody>
-                                        <tr>
-                                            <td>Population</td>
-                                            <td>{country.population}</td>
-                                        </tr>
-                                        <tr>
-                                            <td>Difficulty</td>
-                                            <td>{country.difficulty}</td>
-                                        </tr>
-                                    </tbody></table>
-                                </div>
-                                <button
-                                    onClick={() => this.selectCountryClosePopup(country.name)}
-                                >Start</button>
-                            </div>
-                        </div>
-                    ))}
-                </div>
-            </div>
-        );
-    }
-}
-CountrySelectorPopup.propTypes = {
-    changeCountry: PropTypes.func,
-    closePopup: PropTypes.func,
-};
-
 
 export class CampaignView extends React.Component {
     constructor(props) {
@@ -834,6 +315,7 @@ export class CampaignView extends React.Component {
                 and see how you do in an election.</p>
             </>;
             const altText = 'Nelson Mandela voting in the 1994 South African general election.';
+            const caption = (<>{altText}</>);
             return (
                 <>
                     {this.state.showCountrySelector
@@ -849,14 +331,11 @@ export class CampaignView extends React.Component {
                         setView={() => { this.setState({ showCountrySelector: true }); }}
                         imgFile={'mandela_voting_in_1994.jpg'}
                         altText={altText}
-                        imgCaption={altText}
+                        imgCaption={caption}
                     />
                 </>
             );
         }
-
-
-
         if (!(this.state.populationData && this.state.mapData)) {
             return (<div>Loading!</div>);
         }
